phy: efinix titanium lvds 1000basex: add option for rx termination
Be able to select between internal and external termination for the
RX side of the 1000Base-X PHY. Using the internal ones is not quite right by spec,
because we are using 4 100 Ohm resistors in parallel restulting in 25 Ohm, instead of just one 100 Ohm resistor. Just activating one internal termination resistor lead to a timing problem.

Signed-off-by: Fin Maaß <f.maass@vogl-electronic.com>